refactor(python-bridge): split non-streaming bridge modules
