+++ litellm-rust/crates/python-bridge/src/routes/mod.rs
@@ -1,0 +1,13 @@
+use pyo3::prelude::*;
+
+mod audio_transcription;
+mod chat_completions;
+mod messages;
+mod ocr;
+
+pub(crate) fn register(module: &Bound<'_, PyModule>) -> PyResult<()> {
+    ocr::register(module)?;
+    audio_transcription::register(module)?;
+    messages::register(module)?;
+    chat_completions::register(module)
+}
